High Dimensional SY

High Dimensional SY

HighDimSY(x, ...)

# S3 method for data.frame
HighDimSY(x, group, targetDim, ...)

# S3 method for grouped_df
HighDimSY(x, targetDim, ...)

# S3 method for resample
HighDimSY(x, targetDim, ...)

# S3 method for matrix
HighDimSY(..., targetDim, svdMethod = svd)

Arguments

x

data

...

other options

group

grouping variable

targetDim

target dimension to reduce the data to

svdMethod

svd function used for dimesion reduction by default svd in base is used

Value

list of reduced data, projection matrix, group variable, discrimination function, m matrix.

Examples

HighDimSY(iris, group = Species, targetDim = 1)

#> $reducedData
#> Source: local data frame [150 x 2]
#> Groups: Species [3]
#> 
#>           V1 Species
#>        <dbl>  <fctr>
#> 1  -2.491713  setosa
#> 2  -2.490590  setosa
#> 3  -2.313681  setosa
#> 4  -2.466945  setosa
#> 5  -2.446499  setosa
#> 6  -2.869738  setosa
#> 7  -2.378313  setosa
#> 8  -2.558495  setosa
#> 9  -2.341068  setosa
#> 10 -2.528190  setosa
#> # ... with 140 more rows
#> 
#> $projectionMatrix
#>            [,1]      [,2]       [,3]       [,4]
#> [1,] -0.3245611 0.1275797 -0.8648099 -0.3612303
#> 
#> $group
#> Species
#> 
#> $discrimFunc
#> dqda
#> 
#> $M
#>                init        Sepal.Length Sepal.Width Petal.Length Petal.Width
#> Sepal.Length  0.930  1.582   0.14218367 -0.01403265   0.16654286  0.04544898
#> Sepal.Width  -0.658 -0.454  -0.01403265 -0.04522041   0.07095510  0.03190612
#> Petal.Length  2.798  4.090   0.16654286  0.07095510   0.19065714  0.06703265
#> Petal.Width   1.080  1.780   0.04544898  0.03190612   0.06703265  0.02800000
#>              Sepal.Length  Sepal.Width Petal.Length Petal.Width
#> Sepal.Length  0.280093878 -0.005453061   0.28693469  0.03876327
#> Sepal.Width  -0.005453061 -0.039685714   0.05968163  0.03833061
#> Petal.Length  0.286934694  0.059681633   0.27442857  0.04275510
#> Petal.Width   0.038763265  0.038330612   0.04275510  0.06432653
#> 
#> attr(,"class")
#> [1] "reduced"