Sliced Inverse Regression

Sliced Inverse Regression

SIR(x, ...)

# S3 method for data.frame
SIR(x, group, targetDim, ...)

# S3 method for grouped_df
SIR(x, targetDim, ...)

# S3 method for resample
SIR(x, targetDim, ...)

# S3 method for matrix
SIR(..., targetDim, svdMethod = svd)

Arguments

x

data

...

other options

group

grouping variable

targetDim

target dimension to reduce the data to

svdMethod

svd function used for dimesion reduction by default svd in base is used

Value

list of reduced data, projection matrix, group variable, discrimination function, m matrix.

Examples

SIR(iris, group = Species, targetDim = 1)

#> $reducedData
#> Source: local data frame [150 x 2]
#> Groups: Species [3]
#> 
#>           V1 Species
#>        <dbl>  <fctr>
#> 1  0.6090820  setosa
#> 2  0.5136697  setosa
#> 3  0.5505969  setosa
#> 4  0.4814106  setosa
#> 5  0.6162917  setosa
#> 6  0.5722864  setosa
#> 7  0.5222517  setosa
#> 8  0.5624035  setosa
#> 9  0.4555768  setosa
#> 10 0.5355896  setosa
#> # ... with 140 more rows
#> 
#> $projectionMatrix
#>            [,1]      [,2]       [,3]       [,4]
#> [1,] 0.08480527 0.1569025 -0.2250776 -0.2873743
#> 
#> $group
#> Species
#> 
#> $discrimFunc
#> qda
#> 
#> $M
#>            [,1]        [,2]       [,3]      [,4]
#> [1,] 0.07257170  0.02203172  0.1998074 0.1382330
#> [2,] 0.02203172  0.34123730 -0.1958223 0.1276098
#> [3,] 0.19980741 -0.19582227  0.7467484 0.3149299
#> [4,] 0.13823299  0.12760975  0.3149299 0.2852280
#> 
#> attr(,"class")
#> [1] "reduced"