Classification function

Given a dimension reduced data set a classifier is built using a selected discrimination function.

classify(x, ...)

Arguments

x

data set

...

currently for use of different dicriminant function than the default for the method

Value

discriminant function output plus the projection matrix used, group variable and the method for dimension reduction

Examples

classify(LD(iris, group = Species, targetDim = 1))

#> Call:
#> qda(list(V1 = c(-1.32410293250805, -1.10974377411849, -1.19470975337819, 
#> -1.02229514115045, -1.3373719564055, -1.2091292169819, -1.11357470459669, 
#> -1.21011625174512, -0.969605574019037, -1.15611975706336, -1.3979373623398, 
#> -1.10939859487964, -1.15898893923006, -1.22854842012285, -1.76104226732892, 
#> -1.56148910100143, -1.44284094284116, -1.25371290469614, -1.30853994629703, 
#> -1.29852663302466, -1.17783983961682, -1.19372271861497, -1.48650423146343, 
#> -0.868821281482303, -0.934114800485187, -1.01403277388919, -1.01090826465648, 
#> -1.28681986374357, -1.31083390861061, -1.01942595898374, -1.0061569350863, 
#> -1.15391564692263, -1.56369321114216, -1.64957958949384, -1.08572972925145, 
#> -1.31657227294401, -1.4671103147742, -1.38661712151708, -1.06244739208163, 
#> -1.23126111444545, -1.29099597346062, -0.772305020785596, -1.13127516527719, 
#> -0.904542095630433, -0.994424879353481, -1.01820888360624, -1.31048872937176, 
#> -1.11513695921304, -1.37679249963946, -1.23413029661216, 1.15018840806086, 
#> 1.23059174914514, 1.39299304810052, 1.29770077988025, 1.40553036430074, 
#> 1.37548127881477, 1.40456861598924, 0.769990465795015, 1.2091915593788, 
#> 1.23544191793723, 1.00335701241543, 1.22986004144791, 1.01522026954062, 
#> 1.4437338321572, 0.836360871733969, 1.07275308836519, 1.46857842394337, 
#> 0.943868493417206, 1.6170203405236, 1.00852015689571, 1.72276994047699, 
#> 0.998762170689348, 1.7263455438892, 1.33736766313115, 1.07619749038502, 
#> 1.1283118376633, 1.38856161131746, 1.6689046471876, 1.41841285973981, 
#> 0.648859653926419, 1.00565097472901, 0.87683301545228, 0.967792686111399, 
#> 1.90819824953618, 1.51086814934403, 1.31673345456283, 1.31842691057155, 
#> 1.36225360413685, 1.09408664246027, 1.22887300668469, 1.35778081813426, 
#> 1.3508920140946, 1.060634504174, 0.783259489692461, 1.25575623371843, 
#> 1.06097968341284, 1.16578359782254, 1.11848721578568, 0.588393087214031, 
#> 1.1417695529555, 2.79764197533907, 2.16165805837259, 2.39173669081727, 
#> 2.20885560118753, 2.53056796336637, 2.69500789756932, 1.92944195145843, 
#> 2.40640249443791, 2.37878755970695, 2.56252448270754, 1.91196461429327, 
#> 2.15164474510022, 2.22145955305901, 2.26359279061558, 2.47919431083437, 
#> 2.26113542335897, 2.07372405772427, 2.52736990136348, 3.12758243127838, 
#> 1.95144972324835, 2.38912283571657, 2.12306806205776, 2.73072871171745, 
#> 1.86868785412938, 2.25621861889563, 2.14902190295046, 1.79699089876711, 
#> 1.8077359197367, 2.43329470852072, 1.96021375759256, 2.37320568321764, 
#> 2.16901632594454, 2.50368473633263, 1.73995974702549, 2.07282687513389, 
#> 2.52250343316873, 2.45912633507011, 2.06045503382682, 1.77045285097222, 
#> 2.10747287229608, 2.47778854406212, 2.07296913352545, 2.16165805837259, 
#> 2.52712356134654, 2.53777873014329, 2.20810067698872, 2.06633358660166, 
#> 2.03922031895364, 2.2930253070289, 1.966881508067)), c(1, 1, 
#> 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 
#> 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 
#> 1, 1, 1, 1, 1, 1, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 
#> 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 
#> 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 2, 3, 3, 3, 3, 3, 3, 3, 
#> 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 
#> 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 
#> 3))
#> 
#> Prior probabilities of groups:
#>     setosa versicolor  virginica 
#>  0.3333333  0.3333333  0.3333333 
#> 
#> Group means:
#>                   V1
#> setosa     -1.210844
#> versicolor  1.213978
#> virginica   2.253528