Slow Learning Algorithm

Slow Learning Algorithm

slow_learn(x, ...)

# S3 method for data.frame
slow_learn(x, group, loss = conditional_loss,
  totallossValue, conditionallossValue, method, ...)

# S3 method for grouped_df
slow_learn(x, loss = conditional_loss, totallossValue,
  conditionallossValue, method, ...)

# S3 method for resample
slow_learn(x, loss = conditional_loss, totallossValue,
  conditionallossValue, method, ...)

Arguments

x

data

...

other options

group

grouping variable

loss

loss function

totallossValue

total loss attained with overall reduction

conditionallossValue

loss attained for each dimension reduction

method

method used in slow learn algorithm

Value

list of reduced data, projection matrix, group variable, discrimination function, m matrix and dimension reduction method used.

Examples

slow_learn(iris, group = Species, loss = conditional_loss,
          totallossValue = .1, conditionallossValue = .01, method = SYS)

#> $reducedData
#> Source: local data frame [150 x 2]
#> Groups: Species [3]
#> 
#>          V1 Species
#>       <dbl>  <fctr>
#> 1  2.563807  setosa
#> 2  2.505311  setosa
#> 3  2.354615  setosa
#> 4  2.346557  setosa
#> 5  2.497705  setosa
#> 6  2.583149  setosa
#> 7  2.216109  setosa
#> 8  2.537257  setosa
#> 9  2.237980  setosa
#> 10 2.593641  setosa
#> # ... with 140 more rows
#> 
#> $projectionMatrix
#>           [,1]      [,2]      [,3]       [,4]
#> [1,] 0.3904516 0.2202127 0.7827046 -0.4317724
#> 
#> $group
#> Species
#> 
#> $discrimFunc
#> qda
#> 
#> $energyTotal
#> [1] 0.07331995
#> 
#> attr(,"class")
#> [1] "reduced"